refactor: replace all styles import with @styles alias
