+++ src/screens/Blog/PageTemplate.tsx
@@ -10,8 +10,8 @@ import { BlogContextProvider } from '@screens/Blog/context';
 import { useTwitterScript } from 'hooks/useTwitterScript';
 import LayoutBlog from '@screens/Blog/Layout';
 import { pageTransitionVariants, Tags, Tag } from '@components/Ui';
-import { BlogGlobalStyle } from 'styles/blogPost';
-import { GlobalStyles } from 'styles/index';
+import { BlogGlobalStyle } from '@styles/blogPost';
+import { GlobalStyles } from '@styles/index';
 import { Container, LineDivider } from '@components/Ui';
 import { MenuBar } from '@components/MenuBar';
 import { Frontmatter } from 'types';
