+++ src/styles/blogPost.tsx
@@ -2,7 +2,7 @@ import Typography from 'typography';
 
 import { css, createGlobalStyle } from '@styles/styled';
 import { pandaPrismStyles } from './prims-panda-theme';
-import { FONTS, theme } from 'styles/theme';
+import { FONTS, theme } from '@styles/theme';
 import { headingLinkStyle } from '@components/Ui';
 
 const baseFontSize = '18px';
